Widen serializer revisions to positive i64
Represent revision-gated ASM carriers with one checked signed 64-bit
scalar. Preserve the full positive native lane through parsing, CADIR JSON,
and F3D emission.

Checks: package check with tests, focused regressions, schema-feature tests,
clippy, package tests, source policy, and panic census.

Use --no-verify because the dispatched worker procedure assigns the closing
workspace and hook gates to the integrator.

+++ crates/cadmpeg-asm/src/brep/emit.rs
@@ -3375,7 +3375,7 @@ fn emit_surface_curve_layout<F>(
                 .into_intersection(solved_domain.ok_or("missing procedural curve cache domain")?);
             let curve_tail = cadmpeg_ir::geometry::SurfaceCurveTail::try_new(
                 form.extension,
-                i64::from(form.revision.get()),
+                form.revision.get(),
                 form.cache,
                 form.support_bounds,
                 form.solved_range,
