+++ crates/cadmpeg-codec-f3d/src/writer/generate/native_geometry.rs
@@ -3768,7 +3768,7 @@ fn encode_native_revision_compound_loft(
     native_surface_base(bytes, "spline")?;
     bytes.push(0x0f);
     native_ident(bytes, "cl_loft_spl_sur")?;
-    native_i64(bytes, i64::from(construction.revision().get()));
+    native_i64(bytes, construction.revision().get());
     native_revision_tail_head(
         bytes,
         "compound-loft surface",
@@ -3830,7 +3830,7 @@ fn encode_native_revision_g2_blend(
     native_surface_base(bytes, "spline")?;
     bytes.push(0x0f);
     native_ident(bytes, "g2_blend_spl_sur")?;
-    native_i64(bytes, i64::from(construction.revision().get()));
+    native_i64(bytes, construction.revision().get());
     for parameter in construction.leading_parameters() {
         native_f64(bytes, parameter);
     }
@@ -6493,7 +6493,7 @@ fn native_cache_first_curve_context(
     form: &cadmpeg_ir::geometry::CacheFirstCurveForm,
     solved_cache: Option<&cadmpeg_ir::geometry::nurbs::NurbsCurve>,
 ) -> Result<(), CodecError> {
-    native_i64(bytes, i64::from(form.revision.get()));
+    native_i64(bytes, form.revision.get());
     match &form.cache {
         cadmpeg_ir::geometry::RevisionCacheForm::SolvedCache { fit_tolerance } => {
             native_enum(bytes, 0);
